Add encouragements to mobile view
Based on user testing on the mobile view the encouragement was
contributing to not seeing the question or a clutter screen. The answer
is to remove it from the nav and place under the question. At the same
time remove the done button and reset.
- update the questions and answer data structure in questionsService.js
to include a key/value for a random encouragement.
- update the TestData service in questionsService.js with a function to
inject a random encouragement into the new key/value data
- update the controller in questionsController to call the new TestData
function to update the test Updata
-  update elements in questions.html under the answer element to display
new data.